+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## NEXT VERSION
 
+- chore: remove the use of `Object.values`
+
 ## v1.6.5 (2019-07-11)
 
 - fix: style value `unset` is not supported on IE